backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.5.0

# Utilities
python-dotenv==1.0.0
httpx==0.25.2

# Calendar Integration
caldav==1.3.9

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1